requirements.txt
instagrapi
pillow
moviepy
pydub
yt-dlp
faiss-cpu
streamlit
rootpath
openai
langchain
langchain_openai
langchain_core
langchain_community
beautifulsoup4
selenium
python_dotenv